backend/requirements.txt
tensorflow==2.10.1
Flask==2.2.2
flask-cors
numpy<=1.23.5
Pillow==9.4.0
gunicorn==20.0.4